u64 in addch was throwing an error, so changed it back to u32

+++ src/main.rs
@@ -125,7 +125,7 @@ fn show_add_input(mut todos: &mut Vec<Todo>, screen: &mut i8, window: WINDOW) {
                 }
             } else {
                 todo.push(char::from(c as u8));
-                addch(c as u64);
+                addch(c as u32);
             }
         }
     }
